prediction_service/requirements.txt
fastapi
uvicorn
pillow
numpy
opencv-python
mlflow
ultralytics
logstash-formatter
python-multipart
